refactor: hoist route patterns and path builders to @app/routes
Today the URL pattern lives twice: once in app-routes.tsx as nested
router children ("sets/:setId" + "boards/:boardId") and once in
features/board/paths.ts as absolute strings for generatePath. Renaming
a segment requires editing both.

Move boardPath/boardSetPath next to the route table in src/app/routes.ts
and share segment constants (ROUTE_PATTERNS) between the router config
and the path builders. Feature, loaders, and pages now import the
builders from @app/routes; the board barrel no longer re-exports them.

Addresses §3 P1 from docs/consolidated-action-plan.md. The flat routes
(/, /library, /about) stay as literals — only the parameterized routes
have real duplication worth deduping.
