+++ src/app/app-routes.tsx
@@ -2,6 +2,7 @@ import { AppShell } from "@app/layouts/app-shell";
 import { boardLoader } from "@app/loaders/board-loader";
 import { boardSetIndexLoader } from "@app/loaders/board-set-index-loader";
 import { rootIndexLoader } from "@app/loaders/root-index-loader";
+import { ROUTE_PATTERNS } from "@app/routes";
 import Button from "@mui/material/Button";
 import { m } from "@paraglide/messages.js";
 import { ErrorState } from "@shared/components/error-state";
@@ -43,11 +44,11 @@ const router = createBrowserRouter([
         children: [
           { index: true, loader: rootIndexLoader },
           {
-            path: "sets/:setId",
+            path: ROUTE_PATTERNS.BOARD_SET,
             children: [
               { index: true, loader: boardSetIndexLoader },
               {
-                path: "boards/:boardId",
+                path: ROUTE_PATTERNS.BOARDS,
                 loader: boardLoader,
                 lazy: async () => import("@pages/board-page"),
               },
