+++ src/app/loaders/board-set-index-loader.ts
@@ -1,4 +1,5 @@
-import { boardPath, getBoardSet } from "@features/board";
+import { boardPath } from "@app/routes";
+import { getBoardSet } from "@features/board";
 import { m } from "@paraglide/messages.js";
 import { data, redirect, type LoaderFunctionArgs } from "react-router";
