+++ src/features/board/index.ts
@@ -5,7 +5,6 @@ export { BoardViewer } from "./board-viewer";
 export { importBoardFromUrl } from "./import/from-url";
 export { useImportBoardFiles } from "./import/use-import-board-files";
 export type { BoardRouteParams } from "./navigation/use-board-navigation";
-export { boardPath, boardSetPath } from "./paths";
 export { getBoardSets, removeBoardSet } from "./storage/board-sets-store";
 export type { BoardSetRecord } from "./storage/db";
 export {
