+++ src/pages/library-page.tsx
@@ -1,11 +1,10 @@
 import { PageContainer } from "@app/layouts/page-container";
 import { PageTitle } from "@app/layouts/page-title";
+import { boardPath, boardSetPath } from "@app/routes";
 import {
-  boardPath,
   BoardSetDeleteDialog,
   BoardSetInfoDialog,
   BoardSetList,
-  boardSetPath,
   removeBoardSet,
   useBoardSets,
   useImportBoardFiles,
